+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * [artifact-manager] IAC-799 / Wrong unix file path separators when creating backup path
+* [artifact-manager] IAC-778 / Backup only works with current package version, otherwise throws exception 404 not found
+
 
 ## v2.36.0 - 16 Nov 2023